refactor mutation updates into folder
