+++ frontend/src/apollo/update/index.js
@@ -1,0 +1,4 @@
+export * from './concept'
+export * from './conceptLink'
+export * from './course'
+export * from './courseLink'
