+++ frontend/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import App from './App'
 import { BrowserRouter } from 'react-router-dom'
 import './index.css'
-import client from './apolloClient'
+import client from './apollo/apolloClient'
 import { ApolloProvider } from 'react-apollo-hooks'
 import { isSignedIn } from './lib/authentication'
 import { LoginStateProvider, ErrorStateProvider } from './store'
